requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic[email]==2.7.1
bcrypt==4.1.3
PyJWT==2.8.0
pandas==2.2.2
openpyxl==3.1.2
xlrd==2.0.1
